+++ src/Foldable.ts
@@ -1,9 +1,9 @@
 /**
  * @since 3.0.0
  */
+import { Chain, Chain1, Chain2, Chain2C, Chain3, Chain3C, Chain4 } from './Chain'
 import { flow, pipe } from './function'
 import { HKT, Kind, Kind2, Kind3, Kind4, URIS, URIS2, URIS3, URIS4 } from './HKT'
-import { Monad, Monad1, Monad2, Monad2C, Monad3, Monad3C, Monad4 } from './Monad'
 import { Monoid } from './Monoid'
 
 // -------------------------------------------------------------------------------------
@@ -167,48 +167,51 @@ export function reduceRight<F, G>(
  *
  * @example
  * import { reduceM } from 'fp-ts/Foldable'
- * import { Monad, some } from 'fp-ts/Option'
+ * import { Chain, some } from 'fp-ts/Option'
  * import { make, Foldable } from 'fp-ts/Tree'
  * import { pipe } from 'fp-ts/function'
  *
  * const t = make(1, [make(2, []), make(3, []), make(4, [])])
- * assert.deepStrictEqual(pipe(t, reduceM(Foldable)(Monad)(0, (b, a) => (a > 2 ? some(b + a) : some(b)))), some(7))
+ * assert.deepStrictEqual(pipe(t, reduceM(Foldable)(Chain)(some(0), (b, a) => (a > 2 ? some(b + a) : some(b)))), some(7))
  *
  * @since 3.0.0
  */
 export function reduceM<F extends URIS>(
   F: Foldable1<F>
 ): {
-  <M extends URIS4>(M: Monad4<M>): <B, A, S, R, E>(
-    b: B,
+  <M extends URIS4>(M: Chain4<M>): <S, R, E, B, A>(
+    mb: Kind4<M, S, R, E, B>,
     f: (b: B, a: A) => Kind4<M, S, R, E, B>
   ) => (fa: Kind<F, A>) => Kind4<M, S, R, E, B>
-  <M extends URIS3>(M: Monad3<M>): <B, A, R, E>(
-    b: B,
+  <M extends URIS3>(M: Chain3<M>): <R, E, B, A>(
+    mb: Kind3<M, R, E, B>,
     f: (b: B, a: A) => Kind3<M, R, E, B>
   ) => (fa: Kind<F, A>) => Kind3<M, R, E, B>
-  <M extends URIS3, E>(M: Monad3C<M, E>): <B, A, R>(
-    b: B,
+  <M extends URIS3, E>(M: Chain3C<M, E>): <R, B, A>(
+    mb: Kind3<M, R, E, B>,
     f: (b: B, a: A) => Kind3<M, R, E, B>
   ) => (fa: Kind<F, A>) => Kind3<M, R, E, B>
-  <M extends URIS2>(M: Monad2<M>): <B, A, E>(
-    b: B,
+  <M extends URIS2>(M: Chain2<M>): <E, B, A>(
+    mb: Kind2<M, E, B>,
     f: (b: B, a: A) => Kind2<M, E, B>
   ) => (fa: Kind<F, A>) => Kind2<M, E, B>
-  <M extends URIS2, E>(M: Monad2C<M, E>): <B, A>(
-    b: B,
+  <M extends URIS2, E>(M: Chain2C<M, E>): <B, A>(
+    mb: Kind2<M, E, B>,
     f: (b: B, a: A) => Kind2<M, E, B>
   ) => (fa: Kind<F, A>) => Kind2<M, E, B>
-  <M extends URIS>(M: Monad1<M>): <B, A>(b: B, f: (b: B, a: A) => Kind<M, B>) => (fa: Kind<F, A>) => Kind<M, B>
+  <M extends URIS>(M: Chain1<M>): <B, A>(
+    mb: Kind<M, B>,
+    f: (b: B, a: A) => Kind<M, B>
+  ) => (fa: Kind<F, A>) => Kind<M, B>
 }
 export function reduceM<F>(
   F: Foldable<F>
-): <M>(M: Monad<M>) => <B, A>(b: B, f: (b: B, a: A) => HKT<M, B>) => (fa: HKT<F, A>) => HKT<M, B>
+): <M>(M: Chain<M>) => <B, A>(mb: HKT<M, B>, f: (b: B, a: A) => HKT<M, B>) => (fa: HKT<F, A>) => HKT<M, B>
 export function reduceM<F>(
   F: Foldable<F>
-): <M>(M: Monad<M>) => <B, A>(b: B, f: (b: B, a: A) => HKT<M, B>) => (fa: HKT<F, A>) => HKT<M, B> {
-  return (M) => (b, f) =>
-    F.reduce(M.of(b), (mb, a) =>
+): <M>(M: Chain<M>) => <B, A>(mb: HKT<M, B>, f: (b: B, a: A) => HKT<M, B>) => (fa: HKT<F, A>) => HKT<M, B> {
+  return (M) => (mb, f) =>
+    F.reduce(mb, (mb, a) =>
       pipe(
         mb,
         M.chain((b) => f(b, a))